:new: Isolate everything
