+++ apps/models/package.json
@@ -4,7 +4,6 @@
   "version": "1.0.0",
   "main": "lib/index.js",
   "dependencies": {
-    "helpers": "*",
     "motion-mobx-helpers": "^1.0.2"
   }
 }
