--- apps/views/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "views",
-  "private": true,
-  "version": "1.0.0",
-  "main": "lib/index.js",
-  "babel": {
-    "presets": [
-      "steelbrain"
-    ],
-    "plugins": [
-      [
-        "gloss/transform",
-        {
-          "decoratorName": "view"
-        }
-      ]
-    ]
-  },
-  "dependencies": {
-    "gloss": "^5.0.0",
-    "decorators": "*",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2"
-  }
-}
