+++ apps/web/elements/title.js
@@ -1,15 +1,16 @@
-import { view } from 'helpers'
+import React from 'react'
+import { view } from '../decorators'
 
 @view
 export default class Title {
   static defaultProps = {
-    tag: 'h1'
+    tag: 'h1',
   }
 
   render() {
     const { tag, children, ...props } = this.props
     return (
-      React.createElement(tag, { '$title': true, ...props },
+      React.createElement(tag, { $title: true, ...props },
         children
       )
     )
@@ -23,7 +24,7 @@ export default class Title {
       flexFlow: 'row',
       display: 'flex',
       whiteSpace: 'pre',
-    }
+    },
   }
 
   static theme = {
@@ -31,13 +32,13 @@ export default class Title {
       const reduce = 1 / tag.slice(1)
       const fontSize = +size || 20 + reduce * 20
       return {
-          title: {
+        title: {
           fontSize,
           lineHeight: `${1 + fontSize * 0.06}rem`,
 
           '&:hover': {
-            color: tag === 'a' ? 'red' : 'auto'
-          }
+            color: tag === 'a' ? 'red' : 'auto',
+          },
         },
       }
     },
@@ -49,7 +50,7 @@ export default class Title {
     spaced: {
       title: {
         margin: ['1.5rem', 0],
-      }
-    }
+      },
+    },
   }
 }
