+++ apps/web/elements/views.js
@@ -1,4 +1,4 @@
-import { glossy as $ } from 'helpers'
+import { glossy as $ } from '../decorators'
 
 export const Page = $('section', {
   padding: [0, 10],
@@ -7,11 +7,11 @@ export const Page = $('section', {
 export const Text = $('p', {
   fontSize: 16,
   lineHeight: '1.6rem',
-  display: 'block'
+  display: 'block',
 })
 
 export const Link = $('a', {
-  color: 'red'
+  color: 'red',
 })
 
 export const Quote = $('blockquote', {
@@ -21,5 +21,5 @@ export const Quote = $('blockquote', {
 })
 
 export const Date = $('date', {
-  color: '#999'
+  color: '#999',
 })
