+++ apps/web/package.json
@@ -7,8 +7,7 @@
   },
   "dependencies": {
     "autobind-decorator": "^1.3.4",
-    "gloss": "^5.0.0",
-    "helpers": "*",
+    "gloss": "^5.1.0",
     "lodash": "^4.17.4",
     "mobx": "^3.1.0",
     "mobx-react": "^4.1.0",
@@ -16,6 +15,7 @@
     "morgan": "^1.8.1",
     "motion-hmr": "^1.1.0",
     "motion-mobx-router": "^1.0.2",
+    "motion-view": "*",
     "passport-facebook": "^2.1.1",
     "passport-github": "^1.1.0",
     "passport-google-oauth": "^1.0.0",
